+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 - ~~added code extension support (README currently unused)~~
 - merged DATA and DEV filetypes
-- added **dust** feature to enable optional dusk disk_size calculation
+- ~~added **dust** feature to enable optional dusk disk_size calculation~~
+- replaced **dust** subprocess with filesize crate